Finalized TodoServiceImpl. Added Integration tests for TodoServiceImpl.

+++ src/main/java/gr/evansp/todofullstackappbackend/services/todos/TodoServiceImpl.java
@@ -1,11 +1,20 @@
 package gr.evansp.todofullstackappbackend.services.todos;
 
+import gr.evansp.todofullstackappbackend.exceptions.LogicException;
 import gr.evansp.todofullstackappbackend.exceptions.NotFoundException;
+import gr.evansp.todofullstackappbackend.exceptions.UnauthorizedException;
 import gr.evansp.todofullstackappbackend.models.todos.Todo;
+import gr.evansp.todofullstackappbackend.models.todos.TodoList;
 import gr.evansp.todofullstackappbackend.repositories.todos.TodoRepository;
 import jakarta.transaction.Transactional;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.validation.annotation.Validated;
+
+import java.time.LocalDateTime;
+import java.util.List;
 
 /**
  * Implementation of {@link TodoService}.
@@ -15,35 +24,80 @@ import org.springframework.validation.annotation.Validated;
 @Validated
 public class TodoServiceImpl implements TodoService {
 
+  /**
+   * {@link TodoRepository}.
+   */
   TodoRepository todoRepository;
 
-  public TodoServiceImpl(TodoRepository todoRepository) {
+  /**
+   * {@link TodoListService}
+   */
+  TodoListService todoListService;
+
+  /**
+   * {@link VerifyOwnershipService}
+   */
+  VerifyOwnershipService ownershipService;
+
+  @Autowired
+  public TodoServiceImpl(TodoRepository todoRepository, VerifyOwnershipService ownershipService,
+                         TodoListService todoListService) {
     this.todoRepository = todoRepository;
+    this.ownershipService = ownershipService;
+    this.todoListService = todoListService;
   }
 
   @Override
   @Transactional
   public Todo find(Long id) {
-    return todoRepository.findById(id).orElseThrow(
-        () -> new NotFoundException(NotFoundException.TODO_NOT_FOUND, null)
-    );
+    Todo todo = todoRepository.findById(id).orElseThrow(
+        () -> new NotFoundException(NotFoundException.LIST_NOT_FOUND, null));
+    ownershipService.checkOwnership(todo.getUserId());
+    return todo;
   }
 
   @Override
   @Transactional
   public Todo store(Todo todo) {
+    if (todo.getTodoId() != null) {
+      throw new LogicException(LogicException.ALREADY_EXISTS, new Object[]{todo.getTitle()});
+    }
+    TodoList list = todoListService.find(todo.getTodoListId());
+    ownershipService.checkOwnership(list.getUserId());
+    ownershipService.checkOwnership(todo.getUserId());
+
     return todoRepository.save(todo);
   }
 
   @Override
   @Transactional
   public Todo update(Todo todo) {
+    if (todo.getTodoId() == null) {
+      throw new LogicException(LogicException.DOES_NOT_EXIST, new Object[]{todo.getTitle()});
+    }
+    TodoList list = todoListService.find(todo.getTodoListId());
+    ownershipService.checkOwnership(list.getUserId());
+    ownershipService.checkOwnership(find(todo.getTodoId()).getUserId());
+    ownershipService.checkOwnership(todo.getUserId());
+
+    todo.setLastModified(LocalDateTime.now());
     return todoRepository.save(todo);
   }
 
   @Override
   @Transactional
   public void delete(Todo todo) {
+    ownershipService.checkOwnership(find(todo.getTodoId()).getUserId());
     todoRepository.delete(todo);
+  }
+
+  @Override
+  @Transactional
+  public List<Todo> getAll() {
+    Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+    if (authentication == null) {
+      throw new UnauthorizedException(UnauthorizedException.UNAUTHORIZED, null);
+    }
+    return todoRepository.findByUserId(authentication.getName());
   }
 }
